test(divider): improve divider test coverage
This commit enhances the test coverage for the Divider component.
It introduces new unit tests for prop validators to ensure the align and layout props correctly
handle valid and invalid values.
Additionally, it adds tests for vertical dividers to verify that dashed and align props function as
expected in a vertical layout.

closed #5631