test(divider): improve divider test coverage (#5677)
* test(divider): improve divider test coverage

This commit enhances the test coverage for the Divider component.
It introduces new unit tests for prop validators to ensure the align and layout props correctly
handle valid and invalid values.
Additionally, it adds tests for vertical dividers to verify that dashed and align props function as
expected in a vertical layout.

closed #5631

* test(divider): merged duplicate testing logic and standardized test case formats

- Merged the duplicate testing logic of the divider component

- Standardized the file format

closed #5631

* test: add theme validator

* test(unit): add slot

---------

Co-authored-by: zhangpaopao <zhangpaopao0609@gmail.com>