fix(date-picker): fix incorrect week highlight when switching years  (#5812)
* fix(date-picker): fix incorrect week highlight when switching years in single-week mode

Fixes a bug where, in single-week selection mode, the same week number remained highlighted after
switching to a different year.
This occurred because the highlight logic did not account for the year when comparing the selected
value.
Now, the highlight appears only when both the year and week number match.

* chore: stash changelog [ci skip]

---------

Co-authored-by: baozhangjie <jay.bao@photonpay.com>
Co-authored-by: tdesign-bot <tdesign@tencent.com>